Updated LastQty, OrderQty, MinQty and MaxFloor to be an INT instead of a QTY.
Under Linux these fields marked as QTY field types are compiled to be doubles.
The TSX exchange needs these values to be an integer or it rejects the message.
Since it is defined as QTY which is a double it sends 100.0 instead of 100 and the TSX rejects
the NewOrderSingle messages.